add the first view about client

+++ distrib/models/distributorProfile.py
@@ -4,7 +4,7 @@ from django.core.validators import MaxValueValidator
 
 class DistributorProfile(models.Model):
 	address = models.CharField(blank=True, null=True, max_length=30)
-	phone = models.PositiveIntegerField(help_text="(Solamente dígitos)", validators=[MaxValueValidator(999999999999999)], blank = True, null = True)
+	phone = models.BigIntegerField(help_text="(Solamente dígitos)", validators=[MaxValueValidator(999999999999999)], blank = True, null = True)
 
 	def get_address(self):
 		return self.address
